Update CADability Debugger Visualizer to support Visual Studio 2022
