+++ CADability.DebuggerVisualizers/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MathNet.Numerics.Signed" version="5.0.0" targetFramework="net48" />
+</packages>
