+++ CADability/GeoPoint.cs
@@ -28,9 +28,6 @@ namespace CADability
     [Serializable]
     [DebuggerDisplayAttribute("(x,y,z)={DebugString}")]
     [JsonVersion(serializeAsStruct = true, version = 1)]
-#if DEBUG
-    [System.Diagnostics.DebuggerVisualizer(typeof(GeoPointVisualizer))]
-#endif
     public struct GeoPoint : ISerializable, IJsonSerialize, IExportStep
     {
         /// <summary>
@@ -1113,9 +1110,6 @@ namespace CADability
     [Serializable()]
     [DebuggerDisplayAttribute("(x,y)={DebugString}")]
     [JsonVersion(serializeAsStruct = true, version = 1)]
-#if DEBUG
-    [System.Diagnostics.DebuggerVisualizer(typeof(GeoPoint2DVisualizer))]
-#endif
     public struct GeoPoint2D : ISerializable, IJsonSerialize
     {
         /// <summary>
